requirements.txt
dash>=2.16.0
plotly>=5.20.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
pyjanitor>=0.27.0
gunicorn>=21.2.0
